projects/drag-n-drop/src/lib/drag-n-drop/drag-ndrop.directive.ts
import { Directive, HostListener, Output, EventEmitter, HostBinding } from '@angular/core';

@Directive({
  selector: '[lfDragNDrop]'
})
export class DragNDropDirective {

  @Output() dropFiles = new EventEmitter<any>();

  constructor() { }

  @HostBinding('class.fileover') fileOver: boolean;

  @HostListener('dragover', ['$event']) onDragOver(evt) {
    evt.preventDefault();
    evt.stopPropagation();
    this.fileOver = true;
  }

  @HostListener('dragleave', ['$event']) onDragLeave(evt) {
    evt.preventDefault();
    evt.stopPropagation();
    this.fileOver = false;
  }

  @HostListener('drop', ['$event']) onDrop(evt) {
    evt.preventDefault();
    evt.stopPropagation();
    this.fileOver = false;
    const files = evt.dataTransfer.files;
    if (files && files.length > 0) {
      this.dropFiles.emit(files);
    }
  }
}
